refactor: extract validation & printing
<!-- ps-id: 792a5768-e08b-451e-a6f8-f80886db47aa -->
